pyproject.toml
[tool.poetry]
name = "a_simple_qq_bot"
version = "0.1.0"
description = ""
authors = ["I Love Study <1450069615@qq.com>"]
license = "AGPL-v3"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
graia-ariadne = ">=0.7.0"
prettytable = ">=2.5.0"
pycryptodome = ">=3.12.0"
qrcode = ">=7.3.1"
Pillow = ">=9.0.0"
SQLAlchemy = ">=1.4.29"
matplotlib = ">=3.5.1"
ujson = ">=5.1.0"
graiax-silkcoder = "0.2.0"
graiax-msgparse = ">=0.0.4"
graia-scheduler = ">=0.0.6"
graia-saya = ">=0.0.13"
regex = ">=2021.11.10"
graia-broadcast = ">=0.14.5"
aiohttp = {version = ">=3.7.4.post0", extras = ["speedups"]}
playwright = ">=1.17.2"
aiofile = "^3.7.4"
imageio = "^2.16.1"
prompt_toolkit = "^3.0.28"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
default = false
url = "https://mirrors.aliyun.com/pypi/simple/"

[[tool.poetry.source]]
name = "tuna-tsinghua"
default = false
url = "https://pypi.tuna.tsinghua.edu.cn/simple"